Rather than use one of the google maps packages to manage things, I'm starting with something simpler.
I'll handle all the google maps loading in the normal manner. This lets
me get rid of all the required code for loading the script that the
packages managed in components.

Then rather than using a component architecture to wrap the google maps
functionality, I set it up in parallel and use an adapter to pass
changes between Redux and the Maps.

This has the benefit of avoiding leaky abstractions in the wrappers and
gives finer grained control. It does however create a dual-state system
but I think this will work in my advantage by allowing lower level
control of the maps.