This adds the basic search functionality in.
There is one hacky bit where I have to store state about having just
performed a search in order to avoid dispatching from the render. This
situation was like this.

When we change the exploration map's center we dispatch an event that
changes the center in the store's state.

The context map responds to this and updates its own center.

The center of the exploration map is uncoupled from the store's state
and the source of truth is the map itself. The reason for this is
because we'd have a circular reference where the map changes,
dispatches, which sets the map center...

However, when we perform a search, we also want to set the exploration
maps center.  So in the Exploration map's render, we set center only in
a check for a 'searched' prop. This way, we're not always setting the
center that hasn't yet changed, effectively locking the map position.

There is one last hacky bit. Since in the case of a search, we're
setting the exploration map's center, this would cause a dispatch for
the context map to respond. But we're in a render, so we can't do that.
We must unbind the listener that dispatches to the context map, set the
center and rebind.